@@ -9,7 +9,7 @@
     "build-lib": "ng build ng-magic-iframe && copyfiles -f CHANGELOG.md LICENSE.md README.md dist/ng-magic-iframe",
     "prebuild-lib": "rimraf dist/ng-magic-iframe && npm --prefix ./projects/ng-magic-iframe install ./projects/ng-magic-iframe",
     "build-element": "npm run build-element-es2015 && npm run build-element-es5",
-    "build-element-new": "ng build magic-iframe --prod --extraWebpackConfig ../../webpack.extra.js --output-hashing none --single-bundle true\n",
+    "build-element-new": "ng build magic-iframe --prod --extraWebpackConfig webpack.extra.js --output-hashing none --single-bundle true\n",
     "postbuild-element": "rimraf dist/magic-iframe/{main,polyfills,runtime,scripts}*.js && gzip dist/magic-iframe/*.js && copyfiles -f dist/magic-iframe/* projects/magic-iframe/src/*.html dist/ng-magic-iframe-demo/web-component/",
     "build-element-es2015": "ng build magic-iframe --prod --output-hashing=none",
     "postbuild-element-es2015": "concat ./dist/magic-iframe/runtime.js ./dist/magic-iframe/main.js -o ./dist/magic-iframe/magic-iframe.js",
@@ -31,38 +31,38 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.0.1",
-    "@angular/common": "^8.0.1",
-    "@angular/compiler": "^8.0.1",
-    "@angular/core": "^8.0.1",
-    "@angular/elements": "^8.0.1",
-    "@angular/forms": "^8.0.1",
-    "@angular/http": "^7.2.11",
-    "@angular/platform-browser": "^8.0.1",
-    "@angular/platform-browser-dynamic": "^8.0.1",
-    "@angular/router": "^8.0.1",
+    "@angular/animations": "^8.0.3",
+    "@angular/common": "^8.0.3",
+    "@angular/compiler": "^8.0.3",
+    "@angular/core": "^8.0.3",
+    "@angular/elements": "^8.0.3",
+    "@angular/forms": "^8.0.3",
+    "@angular/http": "^7.2.15",
+    "@angular/platform-browser": "^8.0.3",
+    "@angular/platform-browser-dynamic": "^8.0.3",
+    "@angular/router": "^8.0.3",
     "@ng-bootstrap/ng-bootstrap": "^4.2.1",
-    "@sebgroup/bootstrap": "^3.2.0",
-    "@webcomponents/custom-elements": "^1.2.2",
-    "@webcomponents/webcomponentsjs": "^2.2.7",
-    "core-js": "^2.6.5",
-    "document-register-element": "^1.7.2",
-    "ngx-build-plus": "^7.7.6",
+    "@sebgroup/bootstrap": "^3.4.2",
+    "@webcomponents/custom-elements": "^1.2.4",
+    "@webcomponents/webcomponentsjs": "^2.2.10",
+    "core-js": "^2.6.9",
+    "document-register-element": "^1.13.2",
+    "ngx-build-plus": "^8.1.1",
     "prismjs": "^1.16.0",
     "rxjs": "~6.5.2",
     "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.800.3",
-    "@angular-devkit/build-ng-packagr": "^0.800.3",
-    "@angular/cli": "^8.0.3",
-    "@angular/compiler-cli": "^8.0.1",
-    "@angular/language-service": "^8.0.1",
-    "@commitlint/cli": "^7.5.2",
-    "@commitlint/config-conventional": "^7.5.0",
+    "@angular-devkit/build-angular": "^0.800.6",
+    "@angular-devkit/build-ng-packagr": "^0.800.6",
+    "@angular/cli": "^8.0.6",
+    "@angular/compiler-cli": "^8.0.3",
+    "@angular/language-service": "^8.0.3",
+    "@commitlint/cli": "^7.6.1",
+    "@commitlint/config-conventional": "^7.6.0",
     "@types/jasmine": "^2.8.16",
     "@types/jasminewd2": "^2.0.6",
-    "@types/node": "^8.10.45",
+    "@types/node": "^8.10.49",
     "codelyzer": "^5.1.0",
     "concat": "^1.0.3",
     "copy": "^0.3.2",
@@ -81,13 +81,13 @@
     "ng-packagr": "^5.3.0",
     "protractor": "^5.4.2",
     "replace": "^1.1.0",
-    "semantic-release": "^15.13.3",
+    "semantic-release": "^15.13.18",
     "travis-deploy-once": "^5.0.11",
     "ts-node": "~7.0.0",
     "tslib": "^1.10.0",
     "tslint": "~5.11.0",
-    "typescript": "^3.4.5",
-    "webpack-bundle-analyzer": "^3.1.0"
+    "typescript": "^3.5.2",
+    "webpack-bundle-analyzer": "^3.3.2"
   },
   "commitlint": {
     "extends": [
